Fix #8 and #7 - Readme and Path
Fix incorrect path issue when installed
Update README to reflect proper usage
