+++ src/run.js
@@ -1,0 +1,7 @@
+#! /usr/bin/env node
+
+const EasyApiFixtures = require('./index');
+
+const easyApiFixtures = new EasyApiFixtures();
+
+easyApiFixtures.run();
